requirements.txt
aiodns==3.5.0
aioquic==1.3.0
anyio==4.11.0
async_timeout==5.0.1
brotli==1.2.0
brotlicffi==1.2.0.0
click==8.3.1
ConfigParser==7.2.0
ctypes_snappy==1.03
docutils==0.22.3
filelock==3.20.0
gunicorn==23.0.0
HTMLParser==0.0.2
httpcore==1.0.9
httpx==0.28.1
ipython==8.31.0
ipywidgets==8.1.8
jnius==1.1.0
kerberos==1.3.1
keyring==25.7.0
mask==1.0.0
nacl==0.0.0
orjson==3.11.4
protobuf==6.33.1
pymongo_auth_aws==1.3.0
pymongocrypt==1.17.0
pyOpenSSL==25.3.0
pytest==9.0.1
redis==7.1.0
service_identity==24.2.0
sniffio==1.3.1
Sphinx==8.2.3
thread==2.0.5
trio==0.32.0
urllib3_secure_extra==0.1.0
uvloop==0.22.1
winkerberos==0.12.2
wmi==1.5.1
xmlrpclib==1.0.1
zstandard==0.25.0
discord.py==2.5.2
